+++ src/lib.rs
@@ -40,11 +40,44 @@
 #[doc(no_inline)]
 pub use nonempty::NonEmpty;
 
-// doc-cfg only works on inline items (or those defined here)
-#[doc(inline)]
-#[cfg_attr(all(doc, not(doctest)), doc(cfg(feature = "derive")))]
-#[cfg(any(feature = "derive", doc))]
-pub use unicase::eq as unicase_eq;
+#[cfg(any(feature = "unicode_eq", all(doc, not(doctest))))]
+use std::borrow::Cow;
+
+#[cfg(any(feature = "unicode_eq", all(doc, not(doctest))))]
+fn nfc_normalize(s: &str) -> Cow<'_, str> {
+    use cjk::is_cjkish_codepoint;
+    use unicode_normalization::{is_nfc_quick, IsNormalized, UnicodeNormalization};
+    match is_nfc_quick(s.chars()) {
+        IsNormalized::Yes if !s.chars().any(is_cjkish_codepoint) => Cow::Borrowed(s),
+        _ => Cow::Owned(s.nfc().cjk_compat_variants().collect()),
+    }
+}
+
+#[cfg_attr(all(doc, not(doctest)), doc(cfg(feature = "unicode_eq")))]
+#[cfg(any(feature = "unicode_eq", all(doc, not(doctest))))]
+/// Überprüfe ob zwei Strings nach Unicode Normalisierung identisch sind,
+/// optional ohne Groß-/Kleinschreibung zu beachten.
+///
+/// Diese Funktion verwendet
+/// [Unicode Normalization Form C](https://docs.rs/unicode-normalization/latest/unicode_normalization/trait.UnicodeNormalization.html#tymethod.nfc)
+/// um beide Strings zu vergleichen.
+///
+/// ## English
+/// Check whether two Strings are identical after unicode normalization,
+/// optionally in a [case-insensitive way](unicase::eq).
+///
+/// This funktion uses
+/// [Unicode Normalization Form C](https://docs.rs/unicode-normalization/latest/unicode_normalization/trait.UnicodeNormalization.html#tymethod.nfc)
+/// to compare both strings.
+pub fn unicode_eq(a: &str, b: &str, case_sensitive: bool) -> bool {
+    let a_nfc = nfc_normalize(a);
+    let b_nfc = nfc_normalize(b);
+    if case_sensitive {
+        a_nfc == b_nfc
+    } else {
+        unicase::eq(&a_nfc, &b_nfc)
+    }
+}
 
 #[macro_export]
 /// Crate Name spezifiziert in Cargo.toml.